ocaml_lwt_interop::domain_executor

Function run_in_ocaml_domain

pub unsafe fn run_in_ocaml_domain<T: Send>(
    handle: &Handle,
    f: impl FnOnce(&Runtime) -> T + UnwindSafe,
) -> T

Runs a closure f within the OCaml domain, ensuring the OCaml runtime lock is acquired.

This function spawns a task onto the executor associated with the given handle, and then uses synchronization to ensure that the closure f is executed while the OCaml runtime lock is held by the current thread, while it is being released by the thread, currently running OCaml domain (and OCaml domain executor).

It is rather slow and better avoided until you absolutely have to call some OCaml code on the other thread and retrieve some result synchronously. If you can await a future, it’s awlays better to spawn a task via Handle::spawn and await it’s result there. Tokio provides a nice [tokio::task::block_in_place] which might help avoid calling run_in_ocaml_domain in some cases.

The closure f receives a reference to the OCaml runtime.

Safety

This is only safe to call on a thread that is registered with OCaml runtime, i.e. from a Tokio task, spawned using global Tokio runtime obtained via tokio_rt.

Panics

Panics if any of the synchronization primitives fail.